+++ arm/CMakeLists.txt
@@ -18,7 +18,7 @@ find_package(libhal-pca REQUIRED CONFIG)
 # List of demo applications
 # To add a new demo to the list, simply add its filename without the .cpp
 # extension to this list.
-set(DEMOS app app_serial)
+set(DEMOS app)
 
 # List of targets in the LPC40 series MCU
 set(TARGETS lpc4078)
