+++ arm/applications/app.cpp
@@ -7,9 +7,9 @@
 #include <libhal-util/serial.hpp>
 #include <libhal-util/steady_clock.hpp>
 
-#include "../implementation/mission-control-handler.hpp"
-#include "../implementation/routers/joint-router.hpp"
-#include "../implementation/rules-engine.hpp"
+#include "../implementation/joint_router.hpp"
+#include "../implementation/mission_control_handler.hpp"
+#include "../implementation/rules_engine.hpp"
 
 #include "../hardware_map.hpp"
 #include "src/util.hpp"
@@ -38,8 +38,7 @@ hal::status application(arm::hardware_map& p_map)
     "R0Bot1cs3250",
     hal::create_timeout(counter, 10s).value());
 
-  while (true)
-  {
+  while (true) {
     wifi_result = hal::esp8266::at::wlan_client::create(
       esp,
       "SJSU Robotics 2.4GHz",
@@ -52,7 +51,6 @@ hal::status application(arm::hardware_map& p_map)
     HAL_CHECK(hal::write(terminal, "failed to connect to wifi"));
   }
   HAL_CHECK(hal::write(terminal, "ESP created!\n"));
-      
 
   auto wifi = wifi_result.value();
 
@@ -72,31 +70,25 @@ hal::status application(arm::hardware_map& p_map)
 
   auto socket = std::move(socket_result.value());
 
-  HAL_CHECK(hal::write(terminal, "Beginning Can \n"));
-
   auto can_router = hal::can_router::create(can).value();
 
-  auto rotunda_motor = HAL_CHECK(hal::rmd::drc::create(can_router, counter, 8.0, 0x141));
+  auto rotunda_motor =
+    HAL_CHECK(hal::rmd::drc::create(can_router, counter, 8.0, 0x141));
   auto shoulder_motor =
     HAL_CHECK(hal::rmd::drc::create(can_router, counter, 8 * 65 / 16, 0x142));
-    HAL_CHECK(hal::write(terminal, "Here 1 \n"));
   auto elbow_motor =
     HAL_CHECK(hal::rmd::drc::create(can_router, counter, 8 * 5 / 2, 0x143));
-    HAL_CHECK(hal::write(terminal, "Here 2 \n"));
   auto left_wrist_motor =
     HAL_CHECK(hal::rmd::drc::create(can_router, counter, 8.0, 0x144));
-    HAL_CHECK(hal::write(terminal, "Here 3 \n"));
   auto right_wrist_motor =
     HAL_CHECK(hal::rmd::drc::create(can_router, counter, 8.0, 0x145));
-    HAL_CHECK(hal::write(terminal, "Here 4 \n"));
 
   auto pca9685 = HAL_CHECK(hal::pca::pca9685::create(i2c, 0b100'0000));
-  HAL_CHECK(hal::write(terminal, "Here 5 \n"));
   auto pwm0 = pca9685.get_pwm_channel<0>();
   HAL_CHECK(pwm0.frequency(50.0_Hz));
   std::string_view json;
 
-  arm::JointRouter arm(rotunda_motor,
+  arm::joint_router arm(rotunda_motor,
                        shoulder_motor,
                        elbow_motor,
                        left_wrist_motor,
@@ -105,8 +97,6 @@ hal::status application(arm::hardware_map& p_map)
 
   arm::mc_commands commands;
   arm::motors_feedback feedback;
-  arm::RulesEngine rules_engine;
-  arm::MissionControlHandler mission_control;
 
   HAL_CHECK(hal::delay(counter, 1000ms));
 
@@ -115,51 +105,37 @@ hal::status application(arm::hardware_map& p_map)
     get_request =
       "GET /arm?HB=0&IO=1 HTTP/1.1\r\n Host: 192.168.1.110:5000/\r\n\r\n";
 
-    hal::print(terminal, "here");
-    auto write_result =
-      socket.write(hal::as_bytes(get_request),
-                   hal::create_timeout(counter, 1000ms).value());
+    auto write_result = socket.write(
+      hal::as_bytes(get_request), hal::create_timeout(counter, 1000ms).value());
 
-    hal::print(terminal, "here1");
     if (!write_result) {
       HAL_CHECK(hal::write(terminal, "Failed:  \n"));
       continue;
     }
     HAL_CHECK(hal::delay(counter, 100ms));
 
-    hal::print(terminal, "here2");
-
     auto received = HAL_CHECK(socket.read(buffer)).data;
     auto result = to_string_view(received);
     auto start = result.find('{');
 
-    hal::print<128>(terminal, "%d", start);
     auto end = result.find('}');
-    hal::print<128>(terminal, "%d", end);
-    if(start != -1 && end != -1) {
-      json = result.substr(start, end - start + 1); 
+    if (start != -1 && end != -1) {
+      json = result.substr(start, end - start + 1);
       std::string json_string(json);
-      HAL_CHECK(hal::write(terminal, json)); 
 
-      HAL_CHECK(hal::write(terminal, json_string));
-      HAL_CHECK(hal::write(terminal, "\r\n\n"));
-
-      commands =
-        HAL_CHECK(mission_control.ParseMissionControlData(json_string, terminal));
+      auto new_commands =
+        arm::parse_mission_control_data(json_string, terminal);
+      if (new_commands) {
+        commands = new_commands.value();
+      }
     }
-    
 
-    commands = rules_engine.ValidateCommands(commands);
+    commands = arm::validate_commands(commands);
 
+    commands.print(terminal);
 
-    commands.Print(terminal);
-
-
-    arm.SetJointArguments(commands);
-
-    HAL_CHECK(hal::write(terminal, "Set joint arguments \n"));
+    arm.move(commands);
   }
 
   return hal::success();
 }
-//http://192.168.137.123:5000/arm
