+++ arm/conanfile.py
@@ -11,7 +11,7 @@ class ArmApplication(ConanFile):
         self.requires("libhal-rmd/[^1.0.0]")
         self.requires("libhal-esp8266/[^1.0.0]")
         self.requires("libhal-util/[^1.0.0]")
-        self.requires("libhal-pca/[^0.4.0]")
+        self.requires("libhal-pca/[^1.0.0]")
         self.tool_requires("gnu-arm-embedded-toolchain/11.3.0")
         self.tool_requires("cmake-arm-embedded/0.1.1")
