+++ drive/applications/magnet_demo.cpp
