+++ drive/applications/rmd_demo.cpp
@@ -8,9 +8,9 @@
 // #include <libhal-util/steady_clock.hpp>
 // #include <libhal/can.hpp>
 
-// #include "../dto/drive-dto.hpp"
+// #include "../dto/drive_dto.hpp"
 // #include "../hardware_map.hpp"
-// #include "../implementation/mission-control-handler.hpp"
+// #include "../implementation/mission_control_handler.hpp"
 // #include "src/util.hpp"
 
 // hal::status application(drive::hardware_map& p_map)
@@ -22,8 +22,8 @@
 //   auto& can = *p_map.can;
 //   auto& terminal = *p_map.terminal;
 //   auto& clock = *p_map.steady_clock;
-  // float encoderData = 0.0f;
-  // auto& esp = *p_map.esp;
+// float encoderData = 0.0f;
+// auto& esp = *p_map.esp;
 
 //   // HAL_CHECK(hal::write(terminal, "Starting RMD + WiFi Demo...\n"));
 //   HAL_CHECK(hal::delay(clock, 1s));
@@ -62,7 +62,8 @@
 //   //   });
 
 //   // if (!socket_result) {
-//   //   HAL_CHECK(hal::write(terminal, "TCP Socket couldn't be established\n"));
+//   //   HAL_CHECK(hal::write(terminal, "TCP Socket couldn't be
+//   established\n"));
 //   //   return socket_result.error();
 //   // }
 
@@ -84,7 +85,8 @@
 //   while (true) {
 //     HAL_CHECK(hal::delay(*p_map.steady_clock, 1s))
 
-//     // HAL_CHECK(left_hub_motor.feedback_request(hal::rmd::drc::read::encoder_data));
+//     //
+//     HAL_CHECK(left_hub_motor.feedback_request(hal::rmd::drc::read::encoder_data));
 //     // hal::print<20>(terminal, "%f\n", encoderData);
 //     // HAL_CHECK(hal::write(terminal, encoder));
 //     // buffer.fill('.');
