+++ drive/implementation/command_lerper.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../dto/drive-dto.hpp"
+#include "../dto/drive_dto.hpp"
 
 namespace Drive {
 class command_lerper
