+++ drive/implementation/mission_control_handler.hpp
@@ -5,7 +5,7 @@
 
 #include <string_view>
 
-#include "../dto/drive-dto.hpp"
+#include "../dto/drive_dto.hpp"
 
 namespace Drive {
 // std::string CreateGETRequestParameterWithRoverStatus(drive_commands commands)
