+++ drive/implementation/mode_switcher.hpp
@@ -1,26 +1,27 @@
 #pragma once
 
-#include "../dto/drive-dto.hpp"
-#include "../dto/motor-feedback-dto.hpp"
+#include "../dto/drive_dto.hpp"
+#include "../dto/motor_feedback_dto.hpp"
 
 namespace Drive {
 class mode_switch
 {
 public:
   // Should handle all the logic for switching to new steer mode
-  drive_commands switch_steer_mode(drive_commands commands,
-                                 tri_wheel_router_arguments previous_arguments,
-                                 motor_feedback current_motor_feedback,
-                                 hal::serial& terminal)
+  drive_commands switch_steer_mode(
+    drive_commands commands,
+    tri_wheel_router_arguments previous_arguments,
+    motor_feedback current_motor_feedback,
+    hal::serial& terminal)
   {
     // motor_speeds = HAL_CHECK(tri_wheel.GetMotorFeedback(clock));
     // current_motor_feedback.Print(terminal);
-    bool hubs_stopped((previous_arguments.left.hub.speed >= -0.01f &&
-                       previous_arguments.left.hub.speed <= 0.01f) &&
-                      (previous_arguments.right.hub.speed >= -0.01f &&
-                       previous_arguments.right.hub.speed <= 0.01f) &&
-                      (previous_arguments.back.hub.speed >= -0.01f &&
-                       previous_arguments.back.hub.speed <= 0.01f));
+    bool hubs_stopped((previous_arguments.left.speed >= -0.01f &&
+                       previous_arguments.left.speed <= 0.01f) &&
+                      (previous_arguments.right.speed >= -0.01f &&
+                       previous_arguments.right.speed <= 0.01f) &&
+                      (previous_arguments.back.speed >= -0.01f &&
+                       previous_arguments.back.speed <= 0.01f));
     bool steers_stopped((current_motor_feedback.left_steer_speed >= -0.01f &&
                          current_motor_feedback.left_steer_speed <= 0.01f) &&
                         (current_motor_feedback.right_steer_speed >= -0.01f &&
