+++ science/conanfile.py
@@ -10,7 +10,7 @@ class ScienceApplication(ConanFile):
         self.requires("libhal-lpc/[^1.1.2]")
         self.requires("libhal-esp8266/[^1.0.0]")
         self.requires("libhal-util/[^1.0.0]")
-        self.requires("libhal-pca/[^0.4.0]")
+        self.requires("libhal-pca/[^1.0.0]")
         self.tool_requires("gnu-arm-embedded-toolchain/11.3.0")
         self.tool_requires("cmake-arm-embedded/0.1.1")
